tests/requirements.txt
-r ../requirements.txt
flake8
pytest-mypy
types-PyYAML
mock
types-mock

# remove
ipdb
